Fix: count 변수 타입 수정
float에서 int로 변경

+++ Assets/Scripts/Objects/Item/CloneNumChangeItem.cs
@@ -11,7 +11,7 @@ public enum CloneNumChangeType
 public class CloneNumChangeItem : MonoBehaviour, ICollisionable
 {
     [SerializeField] private CloneNumChangeType type;
-    [SerializeField] private float count;
+    [SerializeField] private int count;
 
     public void OnBulletHit()
     {
